Rename family pages file.
