+++ lib/features/children/family_goal/pages/family_goal_tracker.dart
@@ -4,7 +4,7 @@ import 'package:givt_app/features/children/family_goal/pages/goal_active_widget.
 import 'package:givt_app/features/children/family_goal/pages/goal_completed_widget.dart';
 import 'package:givt_app/features/children/family_goal/pages/no_goal_set_widget.dart';
 import 'package:givt_app/features/children/overview/cubit/family_overview_cubit.dart';
-import 'package:givt_app/features/family/app/pages.dart';
+import 'package:givt_app/features/family/app/family_pages.dart';
 import 'package:givt_app/features/impact_groups/cubit/impact_groups_cubit.dart';
 import 'package:givt_app/features/impact_groups/models/goal.dart';
 import 'package:givt_app/utils/app_theme.dart';
