+++ lib/features/children/overview/pages/family_overview_page.dart
@@ -9,7 +9,7 @@ import 'package:givt_app/features/children/overview/widgets/allowance_warning_di
 import 'package:givt_app/features/children/overview/widgets/children_loading_page.dart';
 import 'package:givt_app/features/children/overview/widgets/family_available_page.dart';
 import 'package:givt_app/features/children/overview/widgets/no_children_page.dart';
-import 'package:givt_app/features/family/app/pages.dart';
+import 'package:givt_app/features/family/app/family_pages.dart';
 import 'package:givt_app/features/family/shared/widgets/top_app_bar.dart';
 import 'package:givt_app/shared/widgets/buttons/leading_back_button.dart';
 import 'package:givt_app/utils/utils.dart';
