+++ lib/features/family/features/coin_flow/screens/search_for_coin_screen.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:givt_app/core/enums/enums.dart';
-import 'package:givt_app/features/family/app/pages.dart';
+import 'package:givt_app/features/family/app/family_pages.dart';
 import 'package:givt_app/features/family/features/coin_flow/cubit/search_coin_cubit.dart';
 import 'package:givt_app/features/family/features/coin_flow/widgets/coin_error_page.dart';
 import 'package:givt_app/features/family/features/coin_flow/widgets/coin_found_page.dart';
